+++ src/loaders/texture_loader.ts
